Rename services module to handlers
